nvmetcli: when kmodpy is not available call kmod binary directly
kmodpy is an unmaintained project. Python is not really good at
backwards compatibility. With the upstream project not keeping up with
python churn using the library is a maintenance burden.

nvmet does not use the python library in any substantial way, it only
loads a module once. This can be easily accomplished without any library
using the modprobe tool directly.

Signed-off-by: Michal Suchanek <msuchanek@suse.de>
[alnovak: use subprocess.run() instead of system(), take into account if
          there are extra arguments in /proc/sys/kernel/modprobe.]
Signed-off-by: Ales Novak <alnovak@suse.com>
Signed-off-by: Daniel Wagner <wagi@kernel.org>